+++ subsys/bluetooth/audio/shell/audio.h
@@ -1,0 +1,25 @@
+/** @file
+ *  @brief Bluetooth audio shell functions
+ *
+ *  This is not to be included by the application.
+ */
+
+/*
+ * Copyright (c) 2023 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#ifndef __AUDIO_H
+#define __AUDIO_H
+
+#include <zephyr/bluetooth/bluetooth.h>
+
+extern struct bt_csip_set_member_svc_inst *svc_inst;
+
+ssize_t audio_ad_data_add(struct bt_data *data, const size_t data_size, const bool discoverable,
+			  const bool connectable);
+ssize_t audio_pa_data_add(struct bt_data *data_array, const size_t data_array_size);
+ssize_t csis_ad_data_add(struct bt_data *data, const size_t data_size, const bool discoverable);
+
+#endif /* __AUDIO_H */
