+++ subsys/bluetooth/audio/shell/csip_set_member.c
@@ -1,22 +1,26 @@
 /**
  * @file
- * @brief Shell APIs for Bluetooth CAP acceptor
+ * @brief Shell APIs for Bluetooth CSIP set member
  *
- * Copyright (c) 2022 Nordic Semiconductor ASA
+ * Copyright (c) 2020 Bose Corporation
+ * Copyright (c) 2021-2022 Nordic Semiconductor ASA
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-#include <stdlib.h>
 
 #include <zephyr/types.h>
+
+#include <zephyr/kernel.h>
+#include <zephyr/types.h>
 #include <zephyr/shell/shell.h>
+#include <stdlib.h>
 #include <zephyr/bluetooth/gatt.h>
 #include <zephyr/bluetooth/bluetooth.h>
-#include <zephyr/bluetooth/audio/cap.h>
-#include "bt.h"
+#include <zephyr/bluetooth/audio/csip.h>
+#include "shell/bt.h"
 
 extern const struct shell *ctx_shell;
-static struct bt_csip_set_member_svc_inst *cap_csip_svc_inst;
+struct bt_csip_set_member_svc_inst *svc_inst;
 static uint8_t sirk_read_rsp = BT_CSIP_READ_SIRK_REQ_RSP_ACCEPT;
 
 static void locked_cb(struct bt_conn *conn,
@@ -57,9 +61,9 @@ static struct bt_csip_set_member_cb csip_set_member_cbs = {
 	.sirk_read_req = sirk_read_req_cb,
 };
 
-static int cmd_cap_acceptor_init(const struct shell *sh, size_t argc,
-				 char **argv)
+static int cm_csip_set_member_register(const struct shell *sh, size_t argc, char **argv)
 {
+	int err;
 	struct bt_csip_set_member_register_param param = {
 		.set_size = 2,
 		.rank = 1,
@@ -69,11 +73,9 @@ static int cmd_cap_acceptor_init(const struct shell *sh, size_t argc,
 			      0x22, 0xfd, 0xa1, 0x21, 0x09, 0x7d, 0x7d, 0x45 },
 		.cb = &csip_set_member_cbs
 	};
-	int err = 0;
 
 	for (size_t argn = 1; argn < argc; argn++) {
 		const char *arg = argv[argn];
-
 		if (strcmp(arg, "size") == 0) {
 			unsigned long set_size;
 
@@ -138,43 +140,37 @@ static int cmd_cap_acceptor_init(const struct shell *sh, size_t argc,
 				      param.set_sirk, sizeof(param.set_sirk));
 			if (len == 0) {
 				shell_error(sh, "Could not parse SIRK");
-
 				return -ENOEXEC;
 			}
 		} else {
 			shell_help(sh);
-
 			return SHELL_CMD_HELP_PRINTED;
 		}
 	}
 
-	err = bt_cap_acceptor_register(&param, &cap_csip_svc_inst);
+	err = bt_csip_set_member_register(&param, &svc_inst);
 	if (err != 0) {
-		shell_error(sh, "Could not register CAS: %d", err);
-
+		shell_error(sh, "Could not register CSIP: %d", err);
 		return err;
 	}
 
 	return 0;
 }
 
-static int cmd_cap_acceptor_print_sirk(const struct shell *sh, size_t argc,
-				       char *argv[])
+static int cm_csip_set_member_print_sirk(const struct shell *sh, size_t argc,
+			       char *argv[])
 {
-	bt_csip_set_member_print_sirk(cap_csip_svc_inst);
-
+	bt_csip_set_member_print_sirk(svc_inst);
 	return 0;
 }
 
-static int cmd_cap_acceptor_lock(const struct shell *sh, size_t argc,
-				 char *argv[])
+static int cm_csip_set_member_lock(const struct shell *sh, size_t argc, char *argv[])
 {
 	int err;
 
-	err = bt_csip_set_member_lock(cap_csip_svc_inst, true, false);
+	err = bt_csip_set_member_lock(svc_inst, true, false);
 	if (err != 0) {
 		shell_error(sh, "Failed to set lock: %d", err);
-
 		return -ENOEXEC;
 	}
 
@@ -183,8 +179,8 @@ static int cmd_cap_acceptor_lock(const struct shell *sh, size_t argc,
 	return 0;
 }
 
-static int cmd_cap_acceptor_release(const struct shell *sh, size_t argc,
-				    char *argv[])
+static int cm_csip_set_member_release(const struct shell *sh, size_t argc,
+			    char *argv[])
 {
 	bool force = false;
 	int err;
@@ -194,16 +190,14 @@ static int cmd_cap_acceptor_release(const struct shell *sh, size_t argc,
 			force = true;
 		} else {
 			shell_error(sh, "Unknown parameter: %s", argv[1]);
-
 			return -ENOEXEC;
 		}
 	}
 
-	err = bt_csip_set_member_lock(cap_csip_svc_inst, false, force);
+	err = bt_csip_set_member_lock(svc_inst, false, force);
 
 	if (err != 0) {
 		shell_error(sh, "Failed to release lock: %d", err);
-
 		return -ENOEXEC;
 	}
 
@@ -212,8 +206,8 @@ static int cmd_cap_acceptor_release(const struct shell *sh, size_t argc,
 	return 0;
 }
 
-static int cmd_cap_acceptor_set_sirk_rsp(const struct shell *sh, size_t argc,
-					 char *argv[])
+static int cm_csip_set_member_set_sirk_rsp(const struct shell *sh, size_t argc,
+				 char *argv[])
 {
 	if (strcmp(argv[1], "accept") == 0) {
 		sirk_read_rsp = BT_CSIP_READ_SIRK_REQ_RSP_ACCEPT;
@@ -231,33 +225,68 @@ static int cmd_cap_acceptor_set_sirk_rsp(const struct shell *sh, size_t argc,
 	return 0;
 }
 
-static int cmd_cap_acceptor(const struct shell *sh, size_t argc, char **argv)
+static int cm_csip_set_member(const struct shell *sh, size_t argc, char **argv)
 {
 	shell_error(sh, "%s unknown parameter: %s", argv[0], argv[1]);
 
 	return -ENOEXEC;
 }
 
-SHELL_STATIC_SUBCMD_SET_CREATE(cap_acceptor_cmds,
-	SHELL_CMD_ARG(init, NULL,
+SHELL_STATIC_SUBCMD_SET_CREATE(csip_set_member_cmds,
+	SHELL_CMD_ARG(register, NULL,
 		      "Initialize the service and register callbacks "
 		      "[size <int>] [rank <int>] [not-lockable] [sirk <data>]",
-		      cmd_cap_acceptor_init, 1, 4),
+		      cm_csip_set_member_register, 1, 4),
 	SHELL_CMD_ARG(lock, NULL,
 		      "Lock the set",
-		      cmd_cap_acceptor_lock, 1, 0),
+		      cm_csip_set_member_lock, 1, 0),
 	SHELL_CMD_ARG(release, NULL,
 		      "Release the set [force]",
-		      cmd_cap_acceptor_release, 1, 1),
+		      cm_csip_set_member_release, 1, 1),
 	SHELL_CMD_ARG(print_sirk, NULL,
 		      "Print the currently used SIRK",
-		      cmd_cap_acceptor_print_sirk, 1, 0),
+		      cm_csip_set_member_print_sirk, 1, 0),
 	SHELL_CMD_ARG(set_sirk_rsp, NULL,
 		      "Set the response used in SIRK requests "
 		      "<accept, accept_enc, reject, oob>",
-		      cmd_cap_acceptor_set_sirk_rsp, 2, 0),
+		      cm_csip_set_member_set_sirk_rsp, 2, 0),
 		      SHELL_SUBCMD_SET_END
 );
 
-SHELL_CMD_ARG_REGISTER(cap_acceptor, &cap_acceptor_cmds, "Bluetooth CAP acceptor shell commands",
-		       cmd_cap_acceptor, 1, 1);
+SHELL_CMD_ARG_REGISTER(csip_set_member, &csip_set_member_cmds,
+		       "Bluetooth CSIP set member shell commands",
+		       cm_csip_set_member, 1, 1);
+
+ssize_t csis_ad_data_add(struct bt_data *data_array, const size_t data_array_size,
+			 const bool discoverable)
+{
+	size_t ad_len = 0;
+
+	/* Advertise RSI in discoverable mode only */
+	if (svc_inst != NULL && discoverable) {
+		static uint8_t ad_rsi[BT_CSIP_RSI_SIZE];
+		int err;
+
+		/* A privacy-enabled Set Member should only advertise RSI values derived
+		 * from a SIRK that is exposed in encrypted form.
+		 */
+		if (IS_ENABLED(CONFIG_BT_PRIVACY) &&
+		    !IS_ENABLED(CONFIG_BT_CSIP_SET_MEMBER_ENC_SIRK_SUPPORT)) {
+			shell_warn(ctx_shell, "RSI derived from unencrypted SIRK");
+		}
+
+		err = bt_csip_set_member_generate_rsi(svc_inst, ad_rsi);
+		if (err != 0) {
+			shell_error(ctx_shell, "Failed to generate RSI (err %d)", err);
+			return err;
+		}
+
+		__ASSERT(data_array_size > ad_len, "No space for AD_RSI");
+		data_array[ad_len].type = BT_DATA_CSIS_RSI;
+		data_array[ad_len].data_len = ARRAY_SIZE(ad_rsi);
+		data_array[ad_len].data = &ad_rsi[0];
+		ad_len++;
+	}
+
+	return ad_len;
+}
