+++ subsys/bluetooth/audio/shell/micp_mic_dev.c
@@ -14,7 +14,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-#include "bt.h"
+#include "shell/bt.h"
 
 static void micp_mic_dev_mute_cb(uint8_t mute)
 {
