+++ subsys/bluetooth/audio/shell/vcp_vol_rend.c
@@ -15,7 +15,7 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "bt.h"
+#include "shell/bt.h"
 
 static struct bt_vcp_included vcp_included;
